fix: sending empty bytes as Input to server on command + usize in dto

+++ src/client/business_objects.rs
@@ -18,7 +18,7 @@ impl From<ServerCommand> for ServerCommandDTO {
 pub enum SessionCommand {
     SplitFocusedWindow(SplitDirection),
     CreateNewWindow,
-    SwitchToWindow(usize),
+    SwitchToWindow(u16),
     FocusPane(FocusDirection),
 }
